Struct freqdist::FrequencyDistribution

pub struct FrequencyDistribution<K, S = DefaultState<SipHasher>> {
    // some fields omitted
}

Methods

impl<K, H = SipHasher, S = DefaultState<SipHasher>> FrequencyDistribution<K, S> where K: Eq + Hash, H: Hasher, S: HashState<Hasher=H>

fn with_capacity_and_hash_state(size: usize, state: S) -> FrequencyDistribution<K, S>

Creates a new FrequencyDistrbution with a hash state and size, where the size is known or can be estimated.

fn with_hash_state(state: S) -> FrequencyDistribution<K, S>

Creates a new FrequencyDistribution with a hash state and default size.

fn keys(&self) -> Keys<K, usize>

Iterator over the keys.

fn iter(&self) -> Iter<K, usize>

Iterator over the key, frequency pairs.

fn iter_non_zero(&self) -> NonZeroKeysIter<K>

Iterator over the non-zero frequency keys.

Example

let existing = vec![
  ("shoes", 1),
  ("scarf", 0),
  ("shirt", 13),
  ("pants", 4)
];

let fdist: FrequencyDistribution<&str> = 
  FromIterator::from_iter(existing.into_iter());
let mut iter = fdist.iter_non_zero();

assert_eq!(*iter.next().unwrap(), "shirt");
assert_eq!(*iter.next().unwrap(), "shoes");
assert_eq!(*iter.next().unwrap(), "pants");
assert!(iter.next().is_none());

fn sum_counts(&self) -> usize

Sum of the total number of items counted thus far.

fn len(&self) -> usize

Returns the number of entries in the distribution

fn get<Q: ?Sized>(&self, k: &Q) -> usize where K: Borrow<Q>, Q: Hash + Eq

Gets the frequency in which the key occurs.

fn clear(&mut self)

Clears the counts of all keys and clears all keys from the distribution.

fn insert(&mut self, k: K)

Updates the frequency of the value found with the key if it already exists. Otherwise, inserts the key sizeo the hashmap, and sets its frequency to 1.

fn remove<Q: ?Sized>(&mut self, k: &Q) where K: Borrow<Q>, Q: Hash + Eq

Removes an item and its associated counts.

impl<K, H = SipHasher, S = DefaultState<SipHasher>> FrequencyDistribution<K, S> where K: Eq + Hash, H: Hasher + Default, S: HashState<Hasher=H> + Default

fn new() -> FrequencyDistribution<K, S>

Creates a new FrequencyDistribution where the size of the HashMap is unknown.

fn with_capacity(size: usize) -> FrequencyDistribution<K, S>

Creates a new FrequencyDistribution where the size of the HashMap is known, or a estimate can be made.